Add abstract media service + simplify video service
